+++ Libplanet.Net/Messages/NetMQMessageCodec.cs
@@ -50,7 +50,7 @@ namespace Libplanet.Net.Messages
 
             // Write headers. (inverse order, version-type-peer-timestamp)
             netMqMessage.Push(timestamp.Ticks);
-            netMqMessage.Push(_codec.Encode(peer.ToBencodex()));
+            netMqMessage.Push(_codec.Encode(peer.Bencoded));
             netMqMessage.Push((int)message.Type);
             netMqMessage.Push(appProtocolVersion.Token);
 
@@ -100,9 +100,7 @@ namespace Libplanet.Net.Messages
             var versionToken = remains[(int)Message.MessageFrame.Version].ConvertToString();
 
             AppProtocolVersion remoteVersion = AppProtocolVersion.FromToken(versionToken);
-            var dictionary =
-                (Bencodex.Types.Dictionary)_codec.Decode(
-                    remains[(int)Message.MessageFrame.Peer].ToByteArray());
+            var dictionary = _codec.Decode(remains[(int)Message.MessageFrame.Peer].ToByteArray());
             BoundPeer remotePeer = new BoundPeer(dictionary);
 
             var type =