+++ Libplanet.Net/Transports/CommunicationFailException.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Runtime.Serialization;
+using Bencodex;
 using Libplanet.Net.Messages;
+using Libplanet.Serialization;
 
 namespace Libplanet.Net.Transports
 {
@@ -12,6 +14,8 @@ namespace Libplanet.Net.Transports
     [Serializable]
     public class CommunicationFailException : Exception
     {
+        private static Codec _codec = new Codec();
+
         public CommunicationFailException(
             string message,
             Message.MessageType messageType,
@@ -36,23 +40,21 @@ namespace Libplanet.Net.Transports
         public CommunicationFailException(SerializationInfo info, StreamingContext context)
             : base(info, context)
         {
-            Peer = info.GetValue(nameof(Peer), typeof(BoundPeer)) is BoundPeer peer
-                ? peer
-                : throw new SerializationException($"{nameof(Peer)} is of an invalid type.");
+            Peer = new BoundPeer(_codec.Decode(info.GetValue<byte[]>(nameof(Peer))));
             MessageType = info.GetValue(nameof(MessageType), typeof(Message.MessageType))
                 is Message.MessageType messageType
                 ? messageType
                 : throw new SerializationException($"{nameof(MessageType)} is of an invalid type.");
         }
 
-        public BoundPeer Peer { get; private set; }
+        public BoundPeer Peer { get; }
 
-        public Message.MessageType MessageType { get; private set; }
+        public Message.MessageType MessageType { get; }
 
         public override void GetObjectData(SerializationInfo info, StreamingContext context)
         {
             base.GetObjectData(info, context);
-            info.AddValue(nameof(Peer), Peer, typeof(BoundPeer));
+            info.AddValue(nameof(Peer), _codec.Encode(Peer.Bencoded));
             info.AddValue(nameof(MessageType), MessageType, typeof(Message.MessageType));
         }
     }